# src/product-service/docs/swagger.yaml
definitions:
  common.AppError:
    properties:
      key:
        type: string
      log:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      filter: {}
      message: {}
      paging: {}
    type: object
...
  models.Warehouses:
    properties:
      location:
        type: string
      warehouse_id:
        type: integer
      warehouse_name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/warehouse:
    post:
      consumes:
      - application/json
      description: Create a new warehouse with the provided parameters
      parameters:
      - description: Warehouse to create
        in: body
        name: warehouse
        required: true
        schema:
          $ref: '#/definitions/models.Warehouses'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Create a new warehouse
      tags:
      - warehouses
  /api/warehouse/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a warehouse by id
      parameters:
      - description: Warehouse ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Deleted a warehouse
      tags:
      - warehouses
    get:
      consumes:
      - application/json
      description: Get a warehouse by id
      parameters:
      - description: Warehouse ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Get a warehouse
      tags:
      - warehouses
    put:
      consumes:
      - application/json
      description: Update an existing warehouse with the provided parameters
      parameters:
      - description: Warehouse to update
        in: body
        name: warehouse
        required: true
        schema:
          $ref: '#/definitions/models.Warehouses'
      - description: Warehouses ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Update an existing warehouse
      tags:
      - warehouses
  /api/warehouses:
    get:
      consumes:
      - application/json
      description: Retrieve all warehouses, with optional paging
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: '-'
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Retrieve a list of warehouses
      tags:
      - warehouses
swagger: "2.0"